Added options to disable the update job and to remove the package
This update allows you to set a parameter to disable the scheduled update
of the package.

It also allow you to set ensure = 'absent' on the module level that will
clean up the crob jobs and all the file under the installation dir. Use
with caution and test it yourself! YMMV (Your Milage May Vary)