Allow custom path for whitelist and allow proxy settings
This update allows you to specify the path to the ispp_scan whitelist file. This can be managed outside the module
(I did not see the added value of doing it via the module)

Also it allows you to specify a webproxy to use which is than set as an environment variabele when issp_scan
is scheduled.